@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using DDCatalogue.Model;
+using DDCatalogue.Contexts;
 using DDCatalogue.Model.Creatures;
 
 namespace DDCatalogue.Controllers
