@@ -1,6 +1,8 @@
 ﻿// <auto-generated />
 using System;
-using DDCatalogue.Model;
+using System.Collections.Generic;
+using DDCatalogue.Contexts;
+using DDCatalogue.Model.Properties;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Migrations;
@@ -12,14 +14,15 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace DDCatalogue.Migrations
 {
     [DbContext(typeof(DDContext))]
-    [Migration("20221114162354_InitialCreate")]
-    partial class InitialCreate
+    [Migration("20230427194309_InitialMigration")]
+    partial class InitialMigration
     {
+        /// <inheritdoc />
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "6.0.0")
+                .HasAnnotation("ProductVersion", "7.0.5")
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.HasPostgresExtension(modelBuilder, "uuid-ossp");
@@ -91,6 +94,8 @@ namespace DDCatalogue.Migrations
                     b.ToTable("Creature");
 
                     b.HasDiscriminator<string>("Discriminator").HasValue("Creature");
+
+                    b.UseTphMappingStrategy();
                 });
 
             modelBuilder.Entity("DDCatalogue.Model.Creatures.Npc", b =>
@@ -142,7 +147,7 @@ namespace DDCatalogue.Migrations
                     b.HasData(
                         new
                         {
-                            Id = new Guid("ada63dfd-4645-49ad-b64d-fc3ccfb642e3"),
+                            Id = new Guid("ce2d5a1c-3b0b-4f15-92bc-4e9580be1d95"),
                             Background = "Raised with his three brothers, Klarg was below average intelligence to say the least. Always playing Kings and Queens, he became obsessed with being the ruler of a country. He forged his own crown, and took over a goblin bandit gang declaring himself as their King.|Klarg has since made a deal with Iarno to steal supplies from Phandalin and to resupply the Redbrands instead.",
                             Beliefs = "[]",
                             Flaws = "[]",
@@ -155,7 +160,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("915c0e65-755d-4a10-9ebc-1d2ef045f5f6"),
+                            Id = new Guid("e5188e4c-6375-4d9e-8ae8-01e97f975bf0"),
                             Background = "",
                             Beliefs = "[]",
                             Flaws = "[]",
@@ -168,7 +173,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("b50fcd71-3fe3-4387-bd2a-8045ec5c3fc9"),
+                            Id = new Guid("99dc4563-34da-4558-a5c7-a39cf1bb296b"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("3829d18f-0c01-4c08-a16f-55d983db578a"),
@@ -182,7 +187,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("60d13f9d-a971-4d4c-aef9-88101fac543f"),
+                            Id = new Guid("3ae33b0e-eff1-42d3-94eb-978151447d3a"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("b3c67d0b-ba6c-4974-b393-5c7d3f1d6644"),
@@ -196,7 +201,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("a54f8ba3-3dfe-47dd-912f-e76f773ab133"),
+                            Id = new Guid("8ad35369-4c60-4bc0-a43a-ecd13fcc5094"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("b439fe6f-dde1-4854-8ea6-a5972a6ea06e"),
@@ -210,7 +215,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("5004c517-1e49-4a93-917a-f6de57262c70"),
+                            Id = new Guid("17286391-d384-4c99-a5d7-25b8b745841b"),
                             Background = "",
                             Beliefs = "[]",
                             Flaws = "[]",
@@ -223,7 +228,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("328e57d4-fe0c-4603-b85f-a54c303b359f"),
+                            Id = new Guid("62375d83-0351-4968-959d-8cd329d279fa"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("5ca1de81-cf25-4c98-a251-1c3fe661457f"),
@@ -237,7 +242,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("b11241c1-f071-4103-a03f-b3dcd806d61c"),
+                            Id = new Guid("6a84b114-50eb-4a8e-8d9f-7af748fb4a7e"),
                             Background = "",
                             Beliefs = "[]",
                             Flaws = "[]",
@@ -250,7 +255,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("f0df09cf-a538-482a-8a08-ffc47b0f27a0"),
+                            Id = new Guid("b3d85d3d-e8b5-4348-9c09-fcf6beb4aa1b"),
                             Background = "",
                             Beliefs = "[]",
                             Flaws = "[]",
@@ -263,7 +268,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("21327109-05b8-44a7-961c-f511101bc7b4"),
+                            Id = new Guid("2af8e063-a51c-4b35-b2aa-afe4c80bcfae"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("6f6bfd5e-a158-487a-b54a-9f27842ce762"),
@@ -277,7 +282,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("436c3a8a-3f57-4393-aad7-2531d1ab2233"),
+                            Id = new Guid("d4368ed6-b706-45d1-9b70-93df51d3d285"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("0110289a-95c6-4385-be5d-741821c8a107"),
@@ -291,7 +296,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("25af5ee4-f7e3-415b-9555-198e75d20a07"),
+                            Id = new Guid("47b3dc26-04c8-48e9-abc0-a0520feca227"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("e54fa7e6-34c0-4ed4-9338-9341162739e5"),
@@ -305,7 +310,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("2ce5db2b-57d8-4078-96bb-ec8f999b2b89"),
+                            Id = new Guid("b1d2b4b7-dab2-4d4b-b757-a47b62ac45e7"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("7042bd21-7239-4866-bcd3-a7a054c05430"),
@@ -319,7 +324,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("b3aa1448-07e6-401b-87ca-a94bdda81f68"),
+                            Id = new Guid("701cf283-f930-42b1-a5df-de4ed109a693"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("4c3bf7ce-dfd6-4894-afa7-f86137003e87"),
@@ -333,7 +338,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("08740102-65fd-47d4-a1c2-eeb07b6e0915"),
+                            Id = new Guid("d00283f9-c643-4ad6-8bff-bd18fb816300"),
                             Background = "",
                             Beliefs = "[]",
                             BuildingId = new Guid("3829d18f-0c01-4c08-a16f-55d983db578a"),
@@ -365,6 +370,8 @@ namespace DDCatalogue.Migrations
                     b.ToTable("Items");
 
                     b.HasDiscriminator<string>("Discriminator").HasValue("Item");
+
+                    b.UseTphMappingStrategy();
                 });
 
             modelBuilder.Entity("DDCatalogue.Model.Joins.BuildingMap", b =>
@@ -678,7 +685,7 @@ namespace DDCatalogue.Migrations
                     b.HasData(
                         new
                         {
-                            Id = new Guid("20d431a5-d40e-4e75-a9c1-ca8ae353e0ed"),
+                            Id = new Guid("7f005891-d2fb-4207-9df9-bb5de4a59bd5"),
                             Name = "Cragmaw Hideout"
                         });
                 });
@@ -704,13 +711,13 @@ namespace DDCatalogue.Migrations
                     b.HasData(
                         new
                         {
-                            Id = new Guid("23f6a61a-bf84-4039-9be4-9eb5192c57c6"),
+                            Id = new Guid("05a500d0-ab28-47dd-8a70-af9befc98589"),
                             Name = "Neverwinter",
                             RegionId = new Guid("050b1c46-0676-420b-872c-143db80d358b")
                         },
                         new
                         {
-                            Id = new Guid("cfbe696c-02af-47e8-9b58-2103a6f04c9b"),
+                            Id = new Guid("8cc5872e-b07b-4134-ae92-a60773901584"),
                             Name = "Thundertree",
                             RegionId = new Guid("050b1c46-0676-420b-872c-143db80d358b")
                         },
@@ -722,13 +729,13 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("70c4e46c-0287-4320-a4dc-6adfd08011b0"),
+                            Id = new Guid("76b7fe93-49cb-45cc-bb11-042999aa0216"),
                             Name = "Conyberry",
                             RegionId = new Guid("050b1c46-0676-420b-872c-143db80d358b")
                         },
                         new
                         {
-                            Id = new Guid("05b8a745-b8aa-4bba-b7c7-507cd352b02e"),
+                            Id = new Guid("c32bcb3c-12db-4b9d-8660-abc61ea9f4b6"),
                             Name = "Old Owl Well",
                             RegionId = new Guid("050b1c46-0676-420b-872c-143db80d358b")
                         },
@@ -740,13 +747,13 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("46a4a3a3-70bb-4782-bfa9-ea13397cb4ac"),
+                            Id = new Guid("3087d4c8-1435-4f90-9a55-b532847190f2"),
                             Name = "Leilon",
                             RegionId = new Guid("050b1c46-0676-420b-872c-143db80d358b")
                         },
                         new
                         {
-                            Id = new Guid("82cb5a0f-a72a-4b3b-adbe-4c540bae2eee"),
+                            Id = new Guid("e6128841-36f5-4acb-8448-a37fbd1806c8"),
                             Name = "Neverwinter",
                             RegionId = new Guid("050b1c46-0676-420b-872c-143db80d358b")
                         },
@@ -804,7 +811,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("a4fa09ec-f479-464b-afd0-1c9e89ff080b"),
+                            Id = new Guid("4fd05e80-0d1f-4497-9886-61d1137c3272"),
                             ImageUrl = "PhanNight.jpg",
                             LocaleId = new Guid("97e40163-3c5c-4f76-b3ce-e6e88ec936ed"),
                             Name = "Phandalin_Night",
@@ -847,8 +854,8 @@ namespace DDCatalogue.Migrations
                 {
                     b.HasBaseType("DDCatalogue.Model.Creatures.Creature");
 
-                    b.Property<string>("Actions")
-                        .HasColumnType("text");
+                    b.Property<List<CreatureAction>>("Actions")
+                        .HasColumnType("jsonb");
 
                     b.Property<double>("ChallengeRating")
                         .HasColumnType("double precision");
@@ -894,7 +901,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":25,\"Measurement\":\"ft\"}]",
                             Strength = 14,
                             Wisdom = 11,
-                            Actions = "[]",
                             ChallengeRating = 1.0,
                             LegendaryActions = "[]",
                             MonsterType = 14,
@@ -923,7 +929,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":30,\"Measurement\":\"ft\"}]",
                             Strength = 15,
                             Wisdom = 11,
-                            Actions = "[{\"name\":\"Morningstar\",\"desc\":\"Melee Weapon Attack: +4 to hit, reach 5 ft., one target. Hit: 11 (2d8 + 2) piercing damage.\",\"attack_bonus\":4,\"damage\":[{\"damage_type\":{\"name\":\"Piercing\"},\"damage_dice\":\"2d8\",\"damage_bonus\":2}]},{\"name\":\"Javelin\",\"desc\":\"Melee or Ranged Weapon Attack: +4 to hit, reach 5 ft. or range 30/120 ft., one target. Hit: 9 (2d6 + 2) piercing damage in melee or 5 (1d6 + 2) piercing damage at range.\",\"attack_bonus\":4,\"damage\":[{\"damage_type\":{\"name\":\"Piercing\"},\"damage_dice\":\"2d6\",\"damage_bonus\":2}]}]",
                             ChallengeRating = 1.0,
                             LegendaryActions = "[]",
                             MonsterType = 14,
@@ -952,7 +957,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":30,\"Measurement\":\"ft\"}]",
                             Strength = 16,
                             Wisdom = 11,
-                            Actions = "[]",
                             ChallengeRating = 3.0,
                             LegendaryActions = "[]",
                             MonsterType = 14,
@@ -981,7 +985,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":30,\"Measurement\":\"ft\"}]",
                             Strength = 8,
                             Wisdom = 8,
-                            Actions = "[]",
                             ChallengeRating = 0.25,
                             LegendaryActions = "[]",
                             MonsterType = 14,
@@ -1010,7 +1013,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":30,\"Measurement\":\"ft\"}]",
                             Strength = 10,
                             Wisdom = 10,
-                            Actions = "[]",
                             ChallengeRating = 0.0,
                             LegendaryActions = "[]",
                             MonsterType = 14,
@@ -1021,7 +1023,7 @@ namespace DDCatalogue.Migrations
                         },
                         new
                         {
-                            Id = new Guid("02108b37-768b-45dc-82ae-968041ee541f"),
+                            Id = new Guid("ab4d8995-b5d1-4e07-aad8-249faee6d16d"),
                             Alignment = 10,
                             ArmorClass = 13,
                             Charisma = 6,
@@ -1039,7 +1041,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":40,\"Measurement\":\"ft\"}]",
                             Strength = 12,
                             Wisdom = 12,
-                            Actions = "[]",
                             ChallengeRating = 0.25,
                             LegendaryActions = "[]",
                             MonsterType = 14,
@@ -1068,7 +1069,6 @@ namespace DDCatalogue.Migrations
                             Speed = "[{\"Name\":\"walk\",\"Value\":25,\"Measurement\":\"ft\"}]",
                             Strength = 13,
                             Wisdom = 12,
-                            Actions = "[]",
                             ChallengeRating = 0.5,
                             LegendaryActions = "[]",
                             MonsterType = 14,
