@@ -1,12 +1,9 @@
 ﻿using System.Globalization;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-using System.Text.Json.Serialization;
+using DDCatalogue.Contexts;
 namespace DDCatalogue.Model
 {
     public class GenericRepository<TEntity> where TEntity : class, IBase
